CC can now be compiled with PointCoordinateType as 'double' (warning: BIN files compatibility is broken however if you do so)

+++ CC/src/CCMiscTools.cpp
@@ -131,37 +131,37 @@ bool CCMiscTools::TriBoxOverlap(PointCoordinateType* boxcenter, PointCoordinateT
 
 	/*  test the 9 tests first (this was faster) */
 	PointCoordinateType rad,fex,fey,fez;		// -NJMP- "d" local variable removed
-	//fex = fabsf(e0[0]);
-	fey = fabsf(e0[1]);
-	fez = fabsf(e0[2]);
+	//fex = fabs(e0[0]);
+	fey = fabs(e0[1]);
+	fez = fabs(e0[2]);
 
 	PointCoordinateType minV,maxV;
 	AXISTEST_X01(e0[2], e0[1], fez, fey);
-	fex = fabsf(e0[0]); //DGM: not necessary before!
+	fex = fabs(e0[0]); //DGM: not necessary before!
 	AXISTEST_Y02(e0[2], e0[0], fez, fex);
 	AXISTEST_Z12(e0[1], e0[0], fey, fex);
 
 	PointCoordinateType e1[3];
 	CCVector3::vsubstract(v2,v1,e1);      /* compute triangle edge 1 */
 
-	//fex = fabsf(e1[0]);
-	fey = fabsf(e1[1]);
-	fez = fabsf(e1[2]);
+	//fex = fabs(e1[0]);
+	fey = fabs(e1[1]);
+	fez = fabs(e1[2]);
 
 	AXISTEST_X01(e1[2], e1[1], fez, fey);
-	fex = fabsf(e1[0]); //DGM: not necessary before!
+	fex = fabs(e1[0]); //DGM: not necessary before!
 	AXISTEST_Y02(e1[2], e1[0], fez, fex);
 	AXISTEST_Z0(e1[1], e1[0], fey, fex);
 
 	PointCoordinateType e2[3];
 	CCVector3::vsubstract(v0,v2,e2);      /* compute triangle edge 2 */
 
-	//fex = fabsf(e2[0]);
-	fey = fabsf(e2[1]);
-	fez = fabsf(e2[2]);
+	//fex = fabs(e2[0]);
+	fey = fabs(e2[1]);
+	fez = fabs(e2[2]);
 
 	AXISTEST_X2(e2[2], e2[1], fez, fey);
-	fex = fabsf(e2[0]); //DGM: not necessary before!
+	fex = fabs(e2[0]); //DGM: not necessary before!
 	AXISTEST_Y1(e2[2], e2[0], fez, fex);
 	AXISTEST_Z12(e2[1], e2[0], fey, fex);
